requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dotenv==1.0.0
APScheduler==3.10.4
reportlab==4.0.7
openpyxl==3.1.2
